xlyz/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-redis==4.12.1
djangorestframework==3.11.0
drf-extensions==0.6.0
psycopg2==2.7.7
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1